+++ Framework/SendbirdUIKit.xcframework/ios-arm64_x86_64-simulator/SendbirdUIKit.framework/Headers/SendbirdUIKit-Swift.h
@@ -2717,8 +2717,14 @@ SWIFT_CLASS("_TtC13SendbirdUIKit18SBUFileMessageCell")
 
 @class NSError;
 
-SWIFT_CLASS("_TtC13SendbirdUIKit13SBUFileViewer")
-@interface SBUFileViewer : SBUBaseViewController <UIScrollViewDelegate>
+/// The <code>SBUBaseViewController</code> that displays file content on <code>FileMessage</code>
+/// <em>Customization Guide</em>
+/// \code
+/// SBUCommonViewControllerSet.FileViewController = MyAppFileViewController.self
+///
+/// \endcode
+SWIFT_CLASS("_TtC13SendbirdUIKit21SBUFileViewController")
+@interface SBUFileViewController : SBUBaseViewController <UIScrollViewDelegate>
 @property (nonatomic, strong) UIBarButtonItem * _Nullable leftBarButton;
 @property (nonatomic, strong) UIBarButtonItem * _Nullable rightBarButton;
 @property (nonatomic, strong) UIView * _Null_unspecified titleView;
@@ -7848,8 +7854,14 @@ SWIFT_CLASS("_TtC13SendbirdUIKit18SBUFileMessageCell")
 
 @class NSError;
 
-SWIFT_CLASS("_TtC13SendbirdUIKit13SBUFileViewer")
-@interface SBUFileViewer : SBUBaseViewController <UIScrollViewDelegate>
+/// The <code>SBUBaseViewController</code> that displays file content on <code>FileMessage</code>
+/// <em>Customization Guide</em>
+/// \code
+/// SBUCommonViewControllerSet.FileViewController = MyAppFileViewController.self
+///
+/// \endcode
+SWIFT_CLASS("_TtC13SendbirdUIKit21SBUFileViewController")
+@interface SBUFileViewController : SBUBaseViewController <UIScrollViewDelegate>
 @property (nonatomic, strong) UIBarButtonItem * _Nullable leftBarButton;
 @property (nonatomic, strong) UIBarButtonItem * _Nullable rightBarButton;
 @property (nonatomic, strong) UIView * _Null_unspecified titleView;