+++ Sample/NotificationService/NotificationService.swift
@@ -7,7 +7,6 @@
 //
 
 import UserNotifications
-import SendbirdChatSDK
 
 class NotificationService: UNNotificationServiceExtension {
     /// > NOTE: How to mark messages as delivered: